input: refactor: remove front-matter package in favor of custom,simple implementation
Front-matter uses js-yaml and esprima for parsing frontmatter blocks from YAML
input. The setup is just too much for our use-case. We need front matter with a couple of properties.

Therefore, decided to get rid of "front-matter" package and write a custom implementation
to parse front-matter as JSON and not YAML.

Results in smaller bundle size and faster execution. Bundle size is not an issue, but
why pay the price for something that is not even needed in first place